Refined the flow a bit
